Move Schedule and Loading to seperate files
