+++ src/client/react/components/container/View.js
@@ -1,30 +1,13 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import createDOMPurify from 'dompurify';
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
 
 import { fetchSchedule } from '../../actions/view';
 import extractSchedule from '../../lib/extractSchedule';
 
-const Loading = () => <div>Loading...</div>;
-
-const Schedule = ({ htmlStr }) => {
-  const DOMPurify = createDOMPurify(window);
-
-  const cleanHTML = DOMPurify.sanitize(htmlStr, {
-    ADD_ATTR: ['rules'],
-  });
-
-  return (
-    // eslint-disable-next-line react/no-danger
-    <div dangerouslySetInnerHTML={{ __html: cleanHTML }} />
-  );
-};
-
-Schedule.propTypes = {
-  htmlStr: PropTypes.string.isRequired,
-};
+import Schedule from '../presentational/Schedule';
+import Loading from '../presentational/Loading';
 
 const View = ({
   schedules,
