+++ src/client/react/components/presentational/Loading.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const Loading = () => <div>Loading...</div>;
+
+export default Loading;
